fixup! feat: Multiple IDE's support
