+++ sources/assets/Stride.Core.Assets.CompilerApp/PackageBuilderApp.cs
@@ -13,13 +13,10 @@ using System.Text;
 using Mono.Options;
 using Stride.Core.Assets.Diagnostics;
 using Stride.Core.BuildEngine;
-using Stride.Core;
 using Stride.Core.Diagnostics;
 using Stride.Core.Yaml;
-using Stride.Core.CodeEditor.VisualStudio;
 using Stride.Assets.Models;
 using Stride.Assets.SpriteFont;
-using Stride.Graphics;
 using Stride.Particles;
 using Stride.Rendering.Materials;
 using Stride.Rendering.ProceduralModels;
