+++ sources/core/Stride.Core.Design/CodeEditor/Rider/RiderLocatorEnvironment.cs
@@ -4,7 +4,7 @@
 using System.Text.Json;
 using JetBrains.Rider.PathLocator;
 
-namespace Stride.Core.CodeEditor.Rider;
+namespace Stride.Core.CodeEditorSupport.Rider;
 
 internal class RiderLocatorEnvironment : IRiderLocatorEnvironment
 {
