+++ sources/core/Stride.Core.Design/Solutions/KnownProjectTypeGuid.cs
@@ -23,17 +23,17 @@
 
 #endregion
 
-namespace Stride.Core.CodeEditor;
+namespace Stride.Core.Solutions;
 
-internal class SolutionFileException : Exception
+public static class KnownProjectTypeGuid
 {
-    public SolutionFileException(string message)
-        : base(message)
-    {
-    }
-
-    public SolutionFileException(string message, Exception innerException)
-        : base(message, innerException)
-    {
-    }
+    public static readonly Guid VisualBasic = new("F184B08F-C81C-45F6-A57F-5ABD9991F28F");
+    public static readonly Guid CSharp = new ("FAE04EC0-301F-11D3-BF4B-00C04F79EFBC");
+    public static readonly Guid JSharp = new("E6FDF86B-F3D1-11D4-8576-0002A516ECE8");
+    public static readonly Guid FSharp = new("F2A71F9B-5D33-465A-A702-920D77279786");
+    public static readonly Guid SolutionFolder = new("2150E333-8FDC-42A3-9474-1A3956D46DE8");
+    public static readonly Guid VisualC = new("8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942");
+    public static readonly Guid Setup = new("54435603-DBB4-11D2-8724-00A0C9A8B90C");
+    public static readonly Guid WebProject = new("E24C65DC-7377-472B-9ABA-BC803B73C61A");
+    public static readonly Guid CSharpNewSystem = new("9A19103F-16F7-4668-BE54-9A1E7A4F7556");
 }
