+++ sources/editor/Stride.GameStudio/ViewModels/GameStudioViewModel.cs
@@ -5,7 +5,6 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Reflection;
-using System.Threading.Tasks;
 using Stride.Core.Assets.Editor.Services;
 using Stride.Core.Assets.Editor.ViewModel;
 using Stride.Core;
@@ -16,12 +15,9 @@ using Stride.Core.MostRecentlyUsedFiles;
 using Stride.Core.Presentation.Commands;
 using Stride.Core.Presentation.Services;
 using Stride.Core.Translation;
-using Stride.Core.CodeEditor.VisualStudio;
 using Stride.Assets.Effect;
 using Stride.Assets.Presentation.ViewModel;
-using Stride.Core.CodeEditor;
-using Stride.Core.CodeEditor.Rider;
-using Stride.Core.CodeEditor.VSCode;
+using Stride.Core.CodeEditorSupport;
 using Stride.GameStudio.Services;
 using Stride.GameStudio.Helpers;
 using Stride.Core.Presentation.ViewModels;
