+++ sources/engine/Stride.Assets/Templates/ProjectTemplateGeneratorHelper.cs
@@ -13,9 +13,7 @@ using Stride.Core.IO;
 using Stride.Core.ProjectTemplating;
 using Stride.Graphics;
 using Stride.Shaders.Parser.Mixins;
-using Stride.Core.CodeEditor.VisualStudio;
 using Stride.Core.Extensions;
-using System.Runtime.InteropServices;
 
 namespace Stride.Assets.Templates
 {
