+++ sources/tools/Stride.FixProjectReferences/FixProjectReference.cs
@@ -7,9 +7,8 @@ using System.Linq;
 using System.Reflection;
 using System.Xml.Linq;
 using Mono.Options;
-using Stride.Core.CodeEditor;
 using Stride.Core.Diagnostics;
-using Stride.Core.CodeEditor.VisualStudio;
+using Stride.Core.Solutions;
 
 namespace Stride.FixProjectReferences
 {
